Rename packages and document SensoAPIInterface
