+++ src/main/java/packages/database_info_manipulation/ReturnMultipleCars.java
@@ -1,7 +1,7 @@
 package packages.database_info_manipulation;
 
-import packages.connect_outer_entity.AutoTraderDBInterface;
-import packages.connect_outer_entity.ConnectAutoTraderDB;
+import packages.connect_api_db.AutoTraderDBInterface;
+import packages.connect_api_db.ConnectAutoTraderDB;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
