+++ src/main/java/packages/database_info_manipulation/ReturnUserInformation.java
@@ -2,8 +2,8 @@ package packages.database_info_manipulation;
 
 import java.sql.*;
 
-import packages.connect_outer_entity.AutoTraderDBInterface;
-import packages.connect_outer_entity.ConnectAutoTraderDB;
+import packages.connect_api_db.AutoTraderDBInterface;
+import packages.connect_api_db.ConnectAutoTraderDB;
 
 import java.sql.ResultSet;
 import java.util.HashMap;
