+++ src/test/java/packages/connect_api_db/ConnectSensoScoreAPITest.java
@@ -1,4 +1,4 @@
-package packages.connect_outer_entity;
+package packages.connect_api_db;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -27,7 +27,6 @@ class ConnectSensoScoreAPITest {
 
     @Test
     public void getReturnInfo() {
-        System.out.println(connector.getReturnInfo());
         assert connector.getReturnInfo() != null;
     }
